+++ models/algorithms/NaiveAlgorithm.py
@@ -1,4 +1,4 @@
-from models.Algorithm import Algorithm
+from models.algorithms.Algorithm import Algorithm
 
 class NaiveAlgorithm(Algorithm):
     def __init__(self):
